+++ docs/source/api_ref_training.rst
@@ -115,4 +115,3 @@ Miscellaneous
 
     get_unmasked_sequence_lengths
     set_seed
-    generation.generate
