+++ tests/torchtune/modules/rlhf/test_generation.py
@@ -8,9 +8,9 @@ import pytest
 
 import torch
 from tests.test_utils import fixed_init_model
+from torchtune.generation._generation import sample
 from torchtune.models.llama2 import llama2
 from torchtune.modules import rlhf
-from torchtune.training.generation._generation import sample
 
 
 class TestGenerateNextTokenWithLogits:
