+++ torchtune/__init__.py
@@ -23,6 +23,6 @@ except ImportError as e:
         """
     ) from e
 
-from torchtune import datasets, models, modules, utils
+from torchtune import datasets, generation, models, modules, utils
 
-__all__ = [datasets, models, modules, utils]
+__all__ = [datasets, models, modules, utils, generation]
