+++ torchtune/training/__init__.py
@@ -66,7 +66,6 @@ from torchtune.training.precision import (
 from torchtune.training.quantization import get_quantizer_mode
 from torchtune.training.seed import set_seed
 
-
 __all__ = [
     "apply_selective_activation_checkpointing",
     "get_dtype",
